Added changes to DTO, repositories and services to create work orders

+++ src/main/java/com/ClinicaDelCalzado_BackEnd/dtos/workOrders/ClientDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class ClientDTO {
 
-    private Integer identification;
+    private Long identification;
     private String name;
     private String cellphone;
 }
